+++ rpcs3/util/atomic.cpp
@@ -269,7 +269,7 @@ static u32 sema_alloc()
 		if (ok)
 		{
 			// Find lowest clear bit
-			const u32 id = group * 64 + utils::cnttz64(~bits, false);
+			const u32 id = group * 64 + static_cast<u32>(utils::cnttz64(~bits, false));
 
 #ifdef USE_POSIX
 			// Initialize semaphore (should be very fast)
